-- on BSD we have this: "int msg_iovlen;". On linux, we have this "size_t msg_iovlen;". So, everybody to uint32_t! NOW! Move! Move! Move!
git-svn-id: https://svn.rtmpd.com/crtmpserver/trunk@459 01689507-d990-df11-9d7d-0017a411d416

+++ sources/thelib/src/protocols/rtp/connectivity/outboundconnectivity.cpp
@@ -285,7 +285,7 @@ bool OutboundConnectivity::FeedDataUDP(msghdr &message, double absoluteTimestamp
 	double rate = isAudio ? _pOutStream->GetCapabilities()->aac._sampleRate : 90000.0;
 	uint32_t ssrc = isAudio ? _pOutStream->AudioSSRC() : _pOutStream->VideoSSRC();
 	uint32_t messageLength = 0;
-	for (int i = 0; i < message.msg_iovlen; i++) {
+	for (uint32_t i = 0; i < (uint32_t) message.msg_iovlen; i++) {
 		messageLength += message.msg_iov[i].iov_len;
 	}
